[hotfix][configuration] Add method for creating restart strategy from configuration
This commit adds method fromConfiguration that allows creating an instance of RestartStrategyConfiguration based on configuration parameters.

It also changes the internal types of RestartStrategyConfigurations to use java.time.Duration instead of Flink's Time. This already solves the equals methods of the configurations. The Flink's Time  class does not have a proper implementation of equals.