[FLINK-14314][runtime] Introduce physical slot ResourceProfile in SlotProfile
The previous resource profile is now named as task resource profile which describes required resources for a logical slot to run the task.
Physical slot resource profile describes the required resources for a physical slot to host this logical task slot. It can be used to allocate a physical slot when no physical slot is available for the requested task slot.