[FLINK-14314][runtime] Introduce physical slot ResourceProfile in SlotProfile
physicalSlotResourceProfile can be used for allocating a physical slot for the task slot. If the task is not in a shared slot, it is the task resource profile. Otherwise it is the slot sharing group resource profile.
The existing ResourceProfile in SlotProfile is renamed as taskResourceProfile for logical slot allocation.